+++ packages/mutable/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @virentia/mutable
 
+## 0.1.2
+
+### Patch Changes
+
+- fix: scoped & remove allSettled
+- Updated dependencies
+  - @virentia/core@0.6.1
+
 ## 0.1.1
 
 ### Patch Changes