+++ packages/vue/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @virentia/vue
 
+## 0.2.7
+
+### Patch Changes
+
+- fix: scoped & remove allSettled
+- Updated dependencies
+  - @virentia/core@0.6.1
+
 ## 0.2.6
 
 ### Patch Changes